Removing earlier versions. Renamed table from details to location.
