+++ services/property_service/alembic/env.py
@@ -50,6 +50,7 @@ def run_migrations_offline() -> None:
         literal_binds=True,
         dialect_opts={"paramstyle": "named"},
         version_table="alembic_version_property_service",
+        version_table_schema="property"
     )
 
     with context.begin_transaction():
@@ -74,7 +75,8 @@ def run_migrations_online() -> None:
         context.configure(
             connection=connection, 
             target_metadata=target_metadata,
-            version_table="alembic_version_property_service"
+            version_table="alembic_property_service_version",
+            version_table_schema="property"
         )
 
         with context.begin_transaction():
