+++ services/property_service/alembic/versions/9bebe894199c_initial_versioning_with_schema_for_.py
@@ -1,0 +1,69 @@
+"""Initial versioning with schema for alembic version table.
+
+Revision ID: 9bebe894199c
+Revises: 
+Create Date: 2025-07-31 16:57:29.532125
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9bebe894199c'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('location',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('address', sa.String(), nullable=True),
+    sa.Column('created_by', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    schema='property'
+    )
+    op.create_index(op.f('ix_property_location_address'), 'location', ['address'], unique=False, schema='property')
+    op.create_index(op.f('ix_property_location_id'), 'location', ['id'], unique=False, schema='property')
+    op.create_index(op.f('ix_property_location_name'), 'location', ['name'], unique=False, schema='property')
+    op.create_table('associations',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('property_id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['property_id'], ['property.location.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    schema='property'
+    )
+    op.create_index(op.f('ix_property_associations_id'), 'associations', ['id'], unique=False, schema='property')
+    op.create_table('rooms',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('property_id', sa.Integer(), nullable=False),
+    sa.Column('room_name', sa.String(), nullable=False),
+    sa.ForeignKeyConstraint(['property_id'], ['property.location.id'], ),
+    sa.PrimaryKeyConstraint('id'),
+    schema='property'
+    )
+    op.create_index(op.f('ix_property_rooms_id'), 'rooms', ['id'], unique=False, schema='property')
+    op.create_index(op.f('ix_property_rooms_room_name'), 'rooms', ['room_name'], unique=False, schema='property')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_property_rooms_room_name'), table_name='rooms', schema='property')
+    op.drop_index(op.f('ix_property_rooms_id'), table_name='rooms', schema='property')
+    op.drop_table('rooms', schema='property')
+    op.drop_index(op.f('ix_property_associations_id'), table_name='associations', schema='property')
+    op.drop_table('associations', schema='property')
+    op.drop_index(op.f('ix_property_location_name'), table_name='location', schema='property')
+    op.drop_index(op.f('ix_property_location_id'), table_name='location', schema='property')
+    op.drop_index(op.f('ix_property_location_address'), table_name='location', schema='property')
+    op.drop_table('location', schema='property')
+    # ### end Alembic commands ###
