--- services/property_service/alembic/versions/daac337c4772_adding_storage.py
@@ -1,73 +1,0 @@
-"""Adding storage.
-
-Revision ID: daac337c4772
-Revises: f3e6d0c100f8
-Create Date: 2025-07-28 22:56:45.699795
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'daac337c4772'
-down_revision: Union[str, Sequence[str], None] = 'f3e6d0c100f8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('alembic_version')
-    op.drop_index(op.f('ix_property_association_id'), table_name='property_association')
-    op.drop_table('property_association')
-    op.drop_index(op.f('ix_user_passwords_id'), table_name='user_passwords')
-    op.drop_table('user_passwords')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('users_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('first_name', sa.VARCHAR(length=50), autoincrement=False, nullable=False),
-    sa.Column('last_name', sa.VARCHAR(length=50), autoincrement=False, nullable=False),
-    sa.Column('email', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.Column('password_hash', postgresql.BYTEA(), autoincrement=False, nullable=False),
-    sa.Column('salt', postgresql.BYTEA(), autoincrement=False, nullable=False),
-    sa.Column('dob', sa.DATE(), autoincrement=False, nullable=True),
-    sa.Column('is_admin', sa.BOOLEAN(), autoincrement=False, nullable=True),
-    sa.Column('is_active', sa.BOOLEAN(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='users_pkey'),
-    sa.UniqueConstraint('email', name='users_email_key', postgresql_include=[], postgresql_nulls_not_distinct=False),
-    postgresql_ignore_search_path=False
-    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('user_passwords',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('email', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.Column('password_str', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name=op.f('user_passwords_user_id_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('user_passwords_pkey'))
-    )
-    op.create_index(op.f('ix_user_passwords_id'), 'user_passwords', ['id'], unique=False)
-    op.create_table('property_association',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('property_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name=op.f('property_association_user_id_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('property_association_pkey'))
-    )
-    op.create_index(op.f('ix_property_association_id'), 'property_association', ['id'], unique=False)
-    op.create_table('alembic_version',
-    sa.Column('version_num', sa.VARCHAR(length=32), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('version_num', name=op.f('alembic_version_pkc'))
-    )
-    # ### end Alembic commands ###
