--- services/property_service/alembic/versions/f3e6d0c100f8_initial_property_migration.py
@@ -1,109 +1,0 @@
-"""Initial property migration.
-
-Revision ID: f3e6d0c100f8
-Revises: 
-Create Date: 2025-07-28 19:55:46.702041
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'f3e6d0c100f8'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('property',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('created_by', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_property_address'), 'property', ['address'], unique=False)
-    op.create_index(op.f('ix_property_id'), 'property', ['id'], unique=False)
-    op.create_index(op.f('ix_property_name'), 'property', ['name'], unique=False)
-    op.create_table('property_associations',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('property_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['property_id'], ['property.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_property_associations_id'), 'property_associations', ['id'], unique=False)
-    op.create_table('property_rooms',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('property_id', sa.Integer(), nullable=False),
-    sa.Column('room_name', sa.String(), nullable=False),
-    sa.ForeignKeyConstraint(['property_id'], ['property.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_property_rooms_id'), 'property_rooms', ['id'], unique=False)
-    op.create_index(op.f('ix_property_rooms_room_name'), 'property_rooms', ['room_name'], unique=False)
-    op.drop_index(op.f('ix_property_association_id'), table_name='property_association')
-    op.drop_table('property_association')
-    op.drop_index(op.f('ix_user_passwords_id'), table_name='user_passwords')
-    op.drop_table('user_passwords')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_table('users')
-    op.drop_table('alembic_version')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('alembic_version',
-    sa.Column('version_num', sa.VARCHAR(length=32), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('version_num', name=op.f('alembic_version_pkc'))
-    )
-    op.create_table('users',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('users_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('first_name', sa.VARCHAR(length=50), autoincrement=False, nullable=False),
-    sa.Column('last_name', sa.VARCHAR(length=50), autoincrement=False, nullable=False),
-    sa.Column('email', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.Column('password_hash', postgresql.BYTEA(), autoincrement=False, nullable=False),
-    sa.Column('salt', postgresql.BYTEA(), autoincrement=False, nullable=False),
-    sa.Column('dob', sa.DATE(), autoincrement=False, nullable=True),
-    sa.Column('is_admin', sa.BOOLEAN(), autoincrement=False, nullable=True),
-    sa.Column('is_active', sa.BOOLEAN(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='users_pkey'),
-    sa.UniqueConstraint('email', name='users_email_key', postgresql_include=[], postgresql_nulls_not_distinct=False),
-    postgresql_ignore_search_path=False
-    )
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('user_passwords',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('email', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.Column('password_str', sa.VARCHAR(length=100), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name=op.f('user_passwords_user_id_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('user_passwords_pkey'))
-    )
-    op.create_index(op.f('ix_user_passwords_id'), 'user_passwords', ['id'], unique=False)
-    op.create_table('property_association',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('property_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name=op.f('property_association_user_id_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('property_association_pkey'))
-    )
-    op.create_index(op.f('ix_property_association_id'), 'property_association', ['id'], unique=False)
-    op.drop_index(op.f('ix_property_rooms_room_name'), table_name='property_rooms')
-    op.drop_index(op.f('ix_property_rooms_id'), table_name='property_rooms')
-    op.drop_table('property_rooms')
-    op.drop_index(op.f('ix_property_associations_id'), table_name='property_associations')
-    op.drop_table('property_associations')
-    op.drop_index(op.f('ix_property_name'), table_name='property')
-    op.drop_index(op.f('ix_property_id'), table_name='property')
-    op.drop_index(op.f('ix_property_address'), table_name='property')
-    op.drop_table('property')
-    # ### end Alembic commands ###
